+++ driver/web/conversions_user.go
@@ -320,6 +320,14 @@ func profileFromDef(item *Def.Profile) model.Profile {
 	if item.PhotoUrl != nil {
 		photoURL = *item.PhotoUrl
 	}
+	var pronunciationURL string
+	if item.PronunciationUrl != nil {
+		pronunciationURL = *item.PronunciationUrl
+	}
+	var pronouns string
+	if item.Pronouns != nil {
+		pronouns = *item.Pronouns
+	}
 	var firstName string
 	if item.FirstName != nil {
 		firstName = *item.FirstName
@@ -348,15 +356,19 @@ func profileFromDef(item *Def.Profile) model.Profile {
 	if item.Country != nil {
 		country = *item.Country
 	}
+	var website string
+	if item.Website != nil {
+		website = *item.Website
+	}
 
 	var unstructuredProperties map[string]interface{}
 	if item.UnstructuredProperties != nil {
 		unstructuredProperties = *item.UnstructuredProperties
 	}
 
-	return model.Profile{PhotoURL: photoURL, FirstName: firstName, LastName: lastName,
-		BirthYear: int16(birthYear), Address: address, ZipCode: zipCode,
-		State: state, Country: country, UnstructuredProperties: unstructuredProperties}
+	return model.Profile{PhotoURL: photoURL, PronunciationURL: pronunciationURL, Pronouns: pronouns, FirstName: firstName,
+		LastName: lastName, BirthYear: int16(birthYear), Address: address, ZipCode: zipCode,
+		State: state, Country: country, Website: website, UnstructuredProperties: unstructuredProperties}
 }
 
 func profileToDef(item *model.Profile) *Def.Profile {
@@ -366,9 +378,10 @@ func profileToDef(item *model.Profile) *Def.Profile {
 
 	itemVal := *item
 	birthYear := int(itemVal.BirthYear)
-	return &Def.Profile{Id: &itemVal.ID, PhotoUrl: &itemVal.PhotoURL, FirstName: &itemVal.FirstName, LastName: &itemVal.LastName,
-		BirthYear: &birthYear, Address: &itemVal.Address, ZipCode: &itemVal.ZipCode, State: &itemVal.State,
-		Country: &itemVal.Country, UnstructuredProperties: &itemVal.UnstructuredProperties}
+	return &Def.Profile{Id: &itemVal.ID, PhotoUrl: &itemVal.PhotoURL, PronunciationUrl: &itemVal.PronunciationURL,
+		Pronouns: &itemVal.Pronouns, FirstName: &itemVal.FirstName, LastName: &itemVal.LastName, BirthYear: &birthYear,
+		Address: &itemVal.Address, ZipCode: &itemVal.ZipCode, State: &itemVal.State, Country: &itemVal.Country, Website: &itemVal.Website,
+		UnstructuredProperties: &itemVal.UnstructuredProperties}
 }
 
 func profileFromDefNullable(item *Def.ProfileNullable) model.Profile {
@@ -380,6 +393,14 @@ func profileFromDefNullable(item *Def.ProfileNullable) model.Profile {
 	if item.PhotoUrl != nil {
 		photoURL = *item.PhotoUrl
 	}
+	var pronunciationURL string
+	if item.PronunciationUrl != nil {
+		pronunciationURL = *item.PronunciationUrl
+	}
+	var pronouns string
+	if item.Pronouns != nil {
+		pronouns = *item.Pronouns
+	}
 	var firstName string
 	if item.FirstName != nil {
 		firstName = *item.FirstName
@@ -408,15 +429,19 @@ func profileFromDefNullable(item *Def.ProfileNullable) model.Profile {
 	if item.Country != nil {
 		country = *item.Country
 	}
+	var website string
+	if item.Website != nil {
+		website = *item.Website
+	}
 
 	var unstructuredProperties map[string]interface{}
 	if item.UnstructuredProperties != nil {
 		unstructuredProperties = *item.UnstructuredProperties
 	}
 
-	return model.Profile{PhotoURL: photoURL, FirstName: firstName, LastName: lastName,
-		BirthYear: int16(birthYear), Address: address, ZipCode: zipCode,
-		State: state, Country: country, UnstructuredProperties: unstructuredProperties}
+	return model.Profile{PhotoURL: photoURL, PronunciationURL: pronunciationURL, Pronouns: pronouns, FirstName: firstName,
+		LastName: lastName, BirthYear: int16(birthYear), Address: address, ZipCode: zipCode,
+		State: state, Country: country, Website: website, UnstructuredProperties: unstructuredProperties}
 }
 
 func privacyToDef(item *model.Privacy) *Def.Privacy {
@@ -425,7 +450,8 @@ func privacyToDef(item *model.Privacy) *Def.Privacy {
 	}
 
 	return &Def.Privacy{
-		Public: &item.Public,
+		Public:          &item.Public,
+		FieldVisibility: &item.FieldVisibility,
 	}
 }
 
@@ -438,8 +464,12 @@ func privacyFromDef(item *Def.Privacy) model.Privacy {
 	if item.Public != nil {
 		public = *item.Public
 	}
+	var fieldVisibility map[string]interface{}
+	if item.FieldVisibility != nil && len(*item.FieldVisibility) > 0 {
+		fieldVisibility = *item.FieldVisibility
+	}
 
-	return model.Privacy{Public: public}
+	return model.Privacy{Public: public, FieldVisibility: fieldVisibility}
 }
 
 func privacyFromDefNullable(item *Def.PrivacyNullable) model.Privacy {
@@ -451,8 +481,12 @@ func privacyFromDefNullable(item *Def.PrivacyNullable) model.Privacy {
 	if item.Public != nil {
 		public = *item.Public
 	}
+	var fieldVisibility map[string]interface{}
+	if item.FieldVisibility != nil && len(*item.FieldVisibility) > 0 {
+		fieldVisibility = *item.FieldVisibility
+	}
 
-	return model.Privacy{Public: public}
+	return model.Privacy{Public: public, FieldVisibility: fieldVisibility}
 }
 
 // MFA