+++ CHANGELOG.md
@@ -24,6 +24,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix email and code sign up
 - Allow update username API to create username
+  
+## [1.45.2] - 2024-18-12
+### Fixed
+- Fix Case-Insensitive Sorting [#741](https://github.com/rokwire/core-building-block/issues/741)
+
+## [1.45.1] - 2024-13-10
+### Fixed
+- GET services/accounts returns unsorted response [#733](https://github.com/rokwire/core-building-block/issues/733)
+
+## [1.45.0] - 2024-12-10
+### Added
+- Public account filtering by unstructured properties [#734](https://github.com/rokwire/core-building-block/issues/734)
+
+## [1.44.1] - 2024-12-03
+### Fixed
+- Fix Privacy.FieldVisibility validation [#729](https://github.com/rokwire/core-building-block/issues/729)
+
+## [1.44.0] - 2024-11-28
+### Added
+- Set up privacy settings for account directory data visibility [#727](https://github.com/rokwire/core-building-block/issues/727)
 
 ## [1.43.0] - 2024-10-21
 ### Added
@@ -474,7 +494,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.42.0...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.44.0...HEAD
+[1.44.0]: https://github.com/rokwire/core-building-block/compare/v1.43.0...v1.44.0
+[1.43.0]: https://github.com/rokwire/core-building-block/compare/v1.42.0...v1.43.0
 [1.42.0]: https://github.com/rokwire/core-building-block/compare/v1.41.0...v1.42.0
 [1.41.0]: https://github.com/rokwire/core-building-block/compare/v1.40.0...v1.41.0
 [1.40.0]: https://github.com/rokwire/core-building-block/compare/v1.39.0...v1.40.0